Make translations discrete and add PointI for that
