+++ geometry/src/lib.rs
@@ -9,6 +9,7 @@ mod cubic_bezier;
 mod flo_curves;
 mod line;
 mod point;
+mod point_i;
 mod projection;
 mod rect;
 mod size;
@@ -23,6 +24,7 @@ pub use color::*;
 pub use cubic_bezier::*;
 pub use line::*;
 pub use point::*;
+pub use point_i::*;
 pub use projection::*;
 pub use rect::*;
 pub use size::*;
